# Enhance Qwen2 with fastsoftmax and bf16 RoPE and cache optimization